Update dot output in defineCnpRegions.
For each region in cnv, defineCnpRegions prints a ".". Previously this
was done using the `cat` function. To make the output friendlier (and
able to be hidden with `message=FALSE` in knitr), use the `message`
function with `appendLF=FALSE` so that all dots are on one line (as
they were with cat).