Change if() to while() in consensusRegion
consensusRegion: Keep halving the threshold until some counts are
beneath the threshold.

consensusRegion checks whether all counts are below a certain threshold.
If so, the threshold is divided in half.  Occasionally, this halved
threshold is still too high and all counts will remain below the
threshold. Essentially, this causes a zero-length vector to be assigned
to the variable index.start. i.e.,

index.start <- min(which(cnt > threshold)

which later results in an indexing error