Change reduced gibbs for theta to use model
Not model.reduced. Have to have chains for reduced gibbs of theta (for
SB, MB, and SBPV). For now, just grab 1:iter(mp) of ptheta.star. Will
need to throw an error if the chains are not long enough. There may be
a better way to do this, so that unnecessary calculations are not
performed (i.e. chains of model are longer than reduced gibbs
iterations)