Adds function reorderSingleBatchChains -- see details
This function reorders all the chains for the SingleBatch model according to the
most frequent ordering of the thetas.  This function was developed when unit
testing the marginal likelihoods for the galaxy data -- I did not find much
value in reordering the chains, but leave the function in case it may prove
useful in other applications.